requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pyTelegramBotAPI==4.28.0
python-dotenv==1.1.1
